+++ choco-solver/src/main/java/org/chocosolver/solver/propagation/NoPropagationEngine.java
@@ -36,7 +36,6 @@
 package org.chocosolver.solver.propagation;
 
 import org.chocosolver.solver.ICause;
-import org.chocosolver.solver.constraints.Constraint;
 import org.chocosolver.solver.constraints.Propagator;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.variables.Variable;
@@ -104,11 +103,15 @@ public enum NoPropagationEngine implements IPropagationEngine {
         }
 
         @Override
-        public void dynamicAddition(Constraint c, boolean permanent) {
+        public void dynamicAddition(boolean permanent, Propagator... ps) {
         }
 
         @Override
-        public void dynamicDeletion(Constraint c) {
+        public void updateInvolvedVariables(Propagator p) {
+        }
+
+        @Override
+        public void dynamicDeletion(Propagator... ps) {
         }
     }
 }