+++ choco-solver/src/main/java/org/chocosolver/solver/propagation/hardcoded/FakeEngine.java
@@ -29,7 +29,6 @@
 package org.chocosolver.solver.propagation.hardcoded;
 
 import org.chocosolver.solver.ICause;
-import org.chocosolver.solver.constraints.Constraint;
 import org.chocosolver.solver.constraints.Propagator;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.exception.SolverException;
@@ -97,11 +96,16 @@ class FakeEngine implements IPropagationEngine {
     }
 
     @Override
-    public void dynamicAddition(Constraint c, boolean permanent) {
+    public void dynamicAddition(boolean permanent, Propagator... ps) {
     }
 
     @Override
-    public void dynamicDeletion(Constraint c) {
+    public void updateInvolvedVariables(Propagator p) {
+
+    }
+
+    @Override
+    public void dynamicDeletion(Propagator... ps) {
 
     }