+++ CHANGES.md
@@ -8,6 +8,7 @@ NEXT MILESTONE
 
 - Add clause_channeling constraint
 - Remove IntVar.wipeOut(...)
+- Enable hot variable addition to propagator
 
 3.3.0 - 04 Dec 2014
 -------------------